+++ lib/fluent/plugin/out_elasticsearch_dynamic.rb
@@ -13,6 +13,8 @@ module Fluent::Plugin
     DYNAMIC_PARAM_NAMES = %W[hosts host port include_timestamp logstash_format logstash_prefix logstash_dateformat time_key utc_index index_name tag_key type_name id_key parent_key routing_key write_operation]
     DYNAMIC_PARAM_SYMBOLS = DYNAMIC_PARAM_NAMES.map { |n| "@#{n}".to_sym }
 
+    RequestInfo = Struct.new(:host, :index)
+
     attr_reader :dynamic_config
 
     def configure(conf)
@@ -42,8 +44,7 @@ module Fluent::Plugin
 
       @_es ||= begin
         @current_config = connection_options[:hosts].clone
-        excon_options = { client_key: @client_key, client_cert: @client_cert, client_key_pass: @client_key_pass }
-        adapter_conf = lambda {|f| f.adapter :excon, excon_options }
+        adapter_conf = lambda {|f| f.adapter @http_backend, @backend_options }
         transport = Elasticsearch::Transport::Transport::HTTP::Faraday.new(connection_options.merge(
                                                                             options: {
                                                                               reload_connections: @reload_connections,
@@ -84,7 +85,7 @@ module Fluent::Plugin
             {
               host:   host_str.split(':')[0],
               port:   (host_str.split(':')[1] || @port).to_i,
-              scheme: @scheme
+              scheme: @scheme.to_s
             }
           else
             # New hosts format expects URLs such as http://logs.foo.com,https://john:pass@logs2.foo.com/elastic
@@ -96,7 +97,7 @@ module Fluent::Plugin
           end
         end.compact
       else
-        [{host: @host, port: @port.to_i, scheme: @scheme}]
+        [{host: @host, port: @port.to_i, scheme: @scheme.to_s}]
       end.each do |host|
         host.merge!(user: @user, password: @password) if !host[:user] && @user
         host.merge!(path: @path) if !host[:path] && @path
@@ -134,10 +135,6 @@ module Fluent::Plugin
 
       chunk.msgpack_each do |time, record|
         next unless record.is_a? Hash
-
-        if @hash_config
-          record = generate_hash_id_key(record)
-        end
 
         begin
           # evaluate all configurations here
@@ -186,7 +183,19 @@ module Fluent::Plugin
           record.merge!(dynamic_conf['tag_key'] => tag)
         end
 
-        meta = {"_index" => target_index, "_type" => dynamic_conf['type_name']}
+        if dynamic_conf['hosts']
+          host = dynamic_conf['hosts']
+        else
+          host = "#{dynamic_conf['host']}:#{dynamic_conf['port']}"
+        end
+
+        if @include_index_in_url
+          key = RequestInfo.new(host, target_index)
+          meta = {"_type" => dynamic_conf['type_name']}
+        else
+          key = RequestInfo.new(host, nil)
+          meta = {"_index" => target_index, "_type" => dynamic_conf['type_name']}
+        end
 
         @meta_config_map.each_pair do |config_name, meta_key|
           if dynamic_conf[config_name] && accessor = record_accessor_create(dynamic_conf[config_name])
@@ -196,30 +205,24 @@ module Fluent::Plugin
           end
         end
 
-        if dynamic_conf['hosts']
-          host = dynamic_conf['hosts']
-        else
-          host = "#{dynamic_conf['host']}:#{dynamic_conf['port']}"
-        end
-
         if @remove_keys
           @remove_keys.each { |key| record.delete(key) }
         end
 
         write_op = dynamic_conf["write_operation"]
-        append_record_to_messages(write_op, meta, headers[write_op], record, bulk_message[host])
+        append_record_to_messages(write_op, meta, headers[write_op], record, bulk_message[key])
       end
 
-      bulk_message.each do |hKey, msgs|
-        send_bulk(msgs, hKey) unless msgs.empty?
+      bulk_message.each do |info, msgs|
+        send_bulk(msgs, info.host, info.index) unless msgs.empty?
         msgs.clear
       end
     end
 
-    def send_bulk(data, host)
+    def send_bulk(data, host, index)
       retries = 0
       begin
-        response = client(host).bulk body: data
+        response = client(host).bulk body: data, index: index
         if response['errors']
           log.error "Could not push log to Elasticsearch: #{response}"
         end
@@ -231,7 +234,7 @@ module Fluent::Plugin
           sleep 2**retries
           retry
         end
-        raise ConnectionFailure, "Could not push logs to Elasticsearch after #{retries} retries. #{e.message}"
+        raise ConnectionRetryFailure, "Could not push logs to Elasticsearch after #{retries} retries. #{e.message}"
       rescue Exception
         @_es = nil if @reconnect_on_error
         raise